+++ backend/services/api-test/src/main/java/io/metersphere/api/dto/request/post/processors/MsPostJSR223Processor.java
@@ -1,4 +1,4 @@
-package io.metersphere.api.dto.jmeter.pre.processors;
+package io.metersphere.api.dto.request.post.processors;
 
 import io.metersphere.plugin.api.annotation.PluginSubType;
 import io.metersphere.plugin.api.dto.TestElementDTO;
@@ -7,9 +7,10 @@ import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-@PluginSubType("MSPreJSR223Processor")
-public class MsPreJSR223Processor extends TestElementDTO {
+@PluginSubType("MSPostJSR223Processor")
+public class MsPostJSR223Processor extends TestElementDTO {
     private String script;
     private String scriptLanguage;
     private Boolean jsrEnable;
+
 }
