+++ backend/services/api-test/src/main/java/io/metersphere/api/dto/request/pre/processors/MsPreJSR223Processor.java
@@ -1,4 +1,4 @@
-package io.metersphere.api.dto.jmeter.post.processors;
+package io.metersphere.api.dto.request.pre.processors;
 
 import io.metersphere.plugin.api.annotation.PluginSubType;
 import io.metersphere.plugin.api.dto.TestElementDTO;
@@ -7,10 +7,9 @@ import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-@PluginSubType("MSPostJSR223Processor")
-public class MsPostJSR223Processor extends TestElementDTO {
+@PluginSubType("MSPreJSR223Processor")
+public class MsPreJSR223Processor extends TestElementDTO {
     private String script;
     private String scriptLanguage;
     private Boolean jsrEnable;
-
 }
