+++ backend/services/api-test/src/main/java/io/metersphere/api/util/ApiDataUtils.java
@@ -11,8 +11,8 @@ import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.databind.jsontype.NamedType;
 import com.fasterxml.jackson.databind.jsontype.impl.StdSubtypeResolver;
 import com.fasterxml.jackson.databind.type.CollectionType;
-import io.metersphere.api.dto.jmeter.post.processors.MsPostJSR223Processor;
-import io.metersphere.api.dto.jmeter.sampler.MsDebugSampler;
+import io.metersphere.api.dto.request.post.processors.MsPostJSR223Processor;
+import io.metersphere.api.dto.request.sampler.MsDebugSampler;
 import io.metersphere.sdk.exception.MSException;
 
 import java.io.IOException;
